refactor: Lines 객체 분리
