+++ app/src/main/java/woowacourse/paint/model/line/Line.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint.model
+package woowacourse.paint.model.line
 
 import android.graphics.Paint
 import android.graphics.Path
